+++ src/separable.jl
@@ -9,12 +9,10 @@
 # This file is part of the `LocalFilters.jl` package licensed under the MIT
 # "Expat" License.
 #
-# Copyright (C) 2018, Éric Thiébaut.
+# Copyright (C) 2018-2020, Éric Thiébaut.
 #
 
 module Separable
-
-using Compat
 
 using ..LocalFilters